move hacked recyclerview into separate package
