+++ app/src/main/res/layout/password_recycler_view.xml
@@ -15,7 +15,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <com.zeapo.pwdstore.ui.FixOnItemTouchDispatchRecyclerView
+        <androidx.recyclerview.widget.FixOnItemTouchDispatchRecyclerView
             android:id="@+id/pass_recycler"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
