+++ spotless.gradle
@@ -13,7 +13,7 @@ spotless {
     }
 
     java {
-        target '**/src/**/*.java'
+        target '**/src/**/com/zeapo/pwdstore/*.java'
         trimTrailingWhitespace()
         licenseHeaderFile rootProject.file('spotless.license')
         removeUnusedImports()
